Move "Bazel Test" GitHub action to a large runner.
"Bazel Test" has been the bottleneck of our builds for a while now. A large part of that is due to the default GitHub runner only having 2 cores and thus only being able to run at most 2 Bazel actions at a time. Things have also gotten worse due to the increasing number of tests in Dagger's codebase and the slowing down of tests due to the move to XProcessing Testing and possibly increased I/O due to golden files, though these issues aren't addressed in this CL.

This CL moves our "Bazel Test" GitHub actions onto the large test runner. This runner now has 16 cores which should allow 8x more tests to run in parallel and we should see a similar improvement in the time it takes these GitHub actions to run.

RELNOTES=N/A
PiperOrigin-RevId: 579881816