Support TYPE_USE @Nullable annotations in Dagger.
When targeting Java 1.8+, TYPE_USE annotations annotate the return type of methods, not the method itself. Similarly, they annotate the type of a parameter, not the parameter itself.

ConfigurationAnnotations now checks the method, the return type of methods, and the type of all elements for a Nullable annotation.

RELNOTES=Adds support for TYPE_USE @Nullable annotations. In rare cases this could be a breaking change originating from the NullableBindingValidation plugin. This could happen if an undetected TYPE_USE @Nullable is used on a module provides method, but @Nullable is not used at injection sites.
PiperOrigin-RevId: 583048123