Update Dagger's GitHub Actions to move the remaining emulator test to run in post-submit only.
The emulator tests have become more flaky recently and are also one of our slowest jobs in GitHub Actions. It should be fairly safe to move these to post-submit given our other non-emulator test coverage.

This CL also updates the release script to use the large-runner. This should be negligible to the overall cost since releases are done so infrequently compared to presubmits.

RELNOTES=N/A
PiperOrigin-RevId: 587014286