Update hjar generators to output valid method bodies for methods and constructors.
This CL updates Dagger's hjar generators to have valid return statements for methods with non-`void` return types and valid super calls for constructors. For example:

```
// Example of hjar generated class with valid constructor and method bodies.
class Foo extends FooSuper {
  Foo() {
    super(null, 0, null);
  }

  String stringMethod() {
    return null;
  }

  int intMethod() {
    return 0;
  }
}
```

Note that this isn't required when compiling with Turbine, but is necessary when compiling the Hjar stubs in Javac.

RELNOTES=N/A
PiperOrigin-RevId: 588456719