Fix error message for an `@Binds @IntoSet` implementation with duplicate bindings.
This CL fixes a case where an `@Binds @IntoSet` that delegates to an implementation with duplicate bindings was throwing an `IllegalArgumentException` rather than reporting an error properly.

An example of the problematic code is shown below:

```java
@Module
interface FooModule {
  @Binds
  @IntoSet
  Foo bindFoo(FooImpl impl);

  @Provides
  static FooImpl provideFooImpl1() { ... }

  @Provides
  static FooImpl provideFooImpl2() { ... }
}
```

The above code now reports the duplicate binding rather than throwing an `IllegalArgumentException`.

RELNOTES=Fixed error message for an `@Binds @IntoSet` implementation with duplicate bindings.
PiperOrigin-RevId: 591976664