Replace `processingEnv.requireTypeElement()` with `DaggerSuperficialValidation.requireTypeElement()`.
I found a case when using `--dagger.pluginsVisitFullBindingGraphs` where `BindingGraphFactory` is triggered from the `ModuleProcessingStep`. In this case, we can't guarantee that the monitoring module has been generated yet so we need to wrap in `DaggerSuperficialValidation.requireTypeElement` to ensure that processing will get delayed if the modules doesn't exist yet.

RELNOTES=N/A
PiperOrigin-RevId: 606341195